+++ Podfile
@@ -64,6 +64,7 @@ def all_pods
     pod 'NotificationBannerSwift', '1.6.3'
     pod 'Nuke', '7.3.2'
     pod 'STRegex', '1.1.0'
+    pod 'Branch', '0.25.5'
 end
 
 def testing_pods
