+++ Stepic/Info.plist
@@ -117,5 +117,12 @@
 	</array>
 	<key>UIViewControllerBasedStatusBarAppearance</key>
 	<false/>
+    <key>branch_key</key>
+    <dict>
+        <key>live</key>
+        <string>key_live_ekt7qHLldFSKQyO2DT3NYellwFfko55Q</string>
+        <key>test</key>
+        <string>key_test_gjtWCOSjnAMKMtV9qJ4YyeelFzced092</string>
+    </dict>
 </dict>
 </plist>
