+++ Stepic/NotificationPermissionStatus.swift
@@ -47,7 +47,7 @@ enum NotificationPermissionStatus {
     @available(iOS 10.0, *)
     init(authorizationStatus: UNAuthorizationStatus) {
         switch authorizationStatus {
-        case .authorized:
+        case .authorized, .provisional:
             self = .authorized
         case .denied:
             self = .denied
